fix: restore frozen MTP plan-stream handoff
