+++ python/sglang/srt/speculative/base_spec_worker.py
@@ -12,6 +12,7 @@ from sglang.srt.model_executor.graph_memory_usage import (
     merge_graph_time_usage,
 )
 from sglang.srt.runtime_context import get_disagg, get_exec, get_memory, get_schedule
+from sglang.srt.speculative.spec_utils import get_plan_stream
 
 if TYPE_CHECKING:
     from sglang.srt.managers.io_struct import (
@@ -62,6 +63,42 @@ class EagleDraftWorkerBase(ABC):
     def __init__(self) -> None:
         self._specialized_graph_memory_usage: dict[str, float] = {}
         self._specialized_graph_time_usage: dict[str, float] = {}
+
+    def _init_handoff_events(self) -> None:
+        self.plan_stream, self.plan_stream_ctx = get_plan_stream(self.device)
+        # Keep handoff events outside captured graphs; per-kind rings avoid
+        # re-recording an event with a pending wait.
+        handoff_ring_size = 8
+        if self.plan_stream is not None:
+            event_cls = torch.get_device_module(self.device).Event
+            self._verify_handoff_events = [
+                event_cls() for _ in range(handoff_ring_size)
+            ]
+            self._draft_extend_handoff_events = [
+                event_cls() for _ in range(handoff_ring_size)
+            ]
+        else:
+            self._verify_handoff_events = []
+            self._draft_extend_handoff_events = []
+        self._verify_handoff_index = 0
+        self._draft_extend_handoff_index = 0
+        self._pending_verify_handoff_event = None
+
+    def _record_handoff_event(self, kind: str, stream):
+        if self.plan_stream is None:
+            return None
+        if kind == "verify":
+            events = self._verify_handoff_events
+            index = self._verify_handoff_index
+            self._verify_handoff_index += 1
+        else:
+            assert kind == "draft_extend"
+            events = self._draft_extend_handoff_events
+            index = self._draft_extend_handoff_index
+            self._draft_extend_handoff_index += 1
+        event = events[index % len(events)]
+        event.record(stream)
+        return event
 
     @abstractmethod
     def draft():
