+++ test/registered/kernels/ops/attention/test_kda_helion.py
@@ -863,7 +863,9 @@ def test_raw_beta_prefill_contract() -> None:
     v = torch.randn(
         batch, tokens, heads, value_dim, device="cuda", dtype=torch.bfloat16
     )
-    gate = torch.randn_like(q) * 0.2
+    # Keep the recurrent decay contractive so the raw-beta check measures the
+    # sigmoid conversion instead of amplifying BF16 round-off exponentially.
+    gate = -torch.rand_like(q) * 0.2
     raw_beta = torch.randn(batch, tokens, heads, device="cuda")
     indices = torch.tensor([3, 1], device="cuda", dtype=torch.int32)
     state = torch.randn(5, heads, value_dim, key_dim, device="cuda") * 0.01
